feat: :sparkles: refactor component imports and update class names for consistency in dashboard sections
