+++ src/app/dashboard/_components/analytics/insights-card.tsx
