+++ src/app/dashboard/_components/exercise/exercise-dialog.tsx
@@ -1,4 +1,4 @@
-import { Dumbbell, Plus } from 'lucide-react'
+import { Dumbbell } from 'lucide-react'
 import { Button } from '~/components/ui/button'
 import {
 	Dialog,
